Start breaking Activities into fragments (breaks MITM for some reason)
The first step in more UI flexibility-- will help with stuff like a
side navigation bar and multi-fragment layouts for tablets.  Unfortunately
this breaks MITM for some reason TBD.  It's the first step.  Will probably
do plugins as well.
